refactor: hooks to return block explorer url
